+++ custom_components/fellow/__init__.py
@@ -50,6 +50,7 @@ PLATFORMS: list[Platform] = [
     Platform.BINARY_SENSOR,
     Platform.BUTTON,
     Platform.NUMBER,
+    Platform.SELECT,
     Platform.UPDATE,
 ]